requirements.txt
# Core Data Science
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Machine Learning
xgboost>=2.0.0
lightgbm>=4.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
scrapy>=2.11.0
selenium>=4.15.0

# APIs
python-dotenv>=1.0.0

# Scheduling
APScheduler>=3.10.0

# Data Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Dashboard
streamlit>=1.28.0

# Development
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.9.0
flake8>=6.1.0
pytest>=7.4.0

# Utilities
tqdm>=4.66.0
python-dateutil>=2.8.0
pytz>=2023.3